refactor: break out attendance builder
